modified src/mcnp_library.cpp to properly read the directory entries
According to an MCNP document,
https://mcnp.lanl.gov/pdf_files/TechReport_2022_LANL_LA-UR-22-30006Rev.1_KuleszaAdamsEtAl.pdf
the lines below "DIRECTORY" (case insensitive) consist of
"seven- to eleven-entry description of each table".
The present implementation, however, seems to assume eleven
entries, and therefore when it encounters a seven-entry line,
it coredumps in e.g. trying to convert a string to double
in mcnp_library.cpp:191

To properly read these variable-entry-number lines, a modification
was made by utilizing std::getline(); the input stream is read
line-by-line and then decomposed.